movement of platform and ball fixed

+++ Animation/Animation.h
@@ -6,10 +6,10 @@ class Animation
 {
 	float speed{};
 	static int scrWidth, scrHeight;
-	void moveRight(float& x, Visible* visible, unsigned dt);
-	void moveLeft(float& x, Visible* visible, unsigned dt);
+	void moveRight(float& x, Visible* visible, float dt);
+	void moveLeft(float& x, Visible* visible, float dt);
 
-	void moveBall(Ball* ball, unsigned dt);
+	void moveBall(Ball* ball, float dt);
 
 	void moveUp(float& y, unsigned int dt, Visible* visible);
 	void moveDown(float& x, unsigned int dt, Visible* visible);
